add test, obsolete storagefill
