+++ Content.Shared/Storage/Components/StorageFillComponent.cs
@@ -3,8 +3,9 @@ using Robust.Shared.GameStates;
 
 namespace Content.Shared.Storage.Components;
 
-// TODO:
-// REPLACE THIS WITH CONTAINERFILL
+// Don't remove before December 2026.
+// A lot of forks still use this for their prototypes.
+[Obsolete("Use ContainerFillComponent or EntityTableContainerFillComponent instead")]
 [RegisterComponent, NetworkedComponent, Access(typeof(SharedStorageSystem))]
 public sealed partial class StorageFillComponent : Component
 {
